added class Team and Course

+++ src/ru/geekbrains/javacore2/lesson1/obstacles/Cross.java
@@ -2,9 +2,9 @@ package ru.geekbrains.javacore2.lesson1.obstacles;
 import ru.geekbrains.javacore2.lesson1.competitors.iChallengable;
 
 public class Cross extends Obstacle {
-    private int distance;
+    private double distance;
 
-    public Cross(int distance) {
+    public Cross(double distance) {
         this.distance = distance;
     }
